+++ app/models/tournament/tournament.rb
@@ -5,7 +5,9 @@ module Tournament
     self.table_name = 'tournaments'
 
     enum :format, { open: 0, swiss: 1, elimination: 2 }
-    enum :state, { draft: 'draft', registration: 'registration', running: 'running', completed: 'completed' }
+    enum :state,
+         { draft: 'draft', registration: 'registration', running: 'running', completed: 'completed',
+           cancelled: 'cancelled' }
 
     belongs_to :creator, class_name: 'User'
     belongs_to :game_system, class_name: 'Game::System'